+++ .changeset/curly-ties-wear.md
@@ -1,0 +1,5 @@
+---
+"@lynx-example/lynx-ui-radio-group": patch
+---
+
+Default `Disabled` example to `lunaris-dark` theme.
