+++ .changeset/skill-lynx-ui-auto.md
@@ -1,0 +1,5 @@
+---
+"@lynx-js/skill-lynx-ui": patch
+---
+
+Update generated lynx-ui skill references.
