+++ .changeset/tender-hats-boil.md
@@ -1,0 +1,5 @@
+---
+"@lynx-example/lynx-ui-feed-list": patch
+---
+
+Unify `Basic` and `Refresh` example styles.
