+++ apps/examples/FeedList/shared/rectangle-card.css
@@ -1,0 +1,43 @@
+@import "@lynx-js/luna-styles/index.css";
+
+.rectangle-card {
+  background-color: var(--canvas);
+  border-color: var(--line);
+  border-width: 1px;
+  border-radius: 16px;
+  margin-left: 16px;
+  margin-right: 16px;
+  margin-bottom: 12px;
+  padding-top: 80px;
+  padding-bottom: 80px;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+}
+
+.rectangle-card__letter {
+  font-size: 224px;
+  font-weight: 600;
+  padding-left: 48px;
+  padding-right: 48px;
+  color: var(--canvas);
+  border-radius: 16px;
+  background: linear-gradient(
+    150deg,
+    var(--primary) 8.42%,
+    var(--primary-2) 81.71%
+  );
+}
+
+.rectangle-card__title {
+  font-size: 18px;
+  font-weight: 600;
+  margin-top: auto;
+  color: var(--content);
+}
+
+.rectangle-card__subtitle {
+  font-size: 12px;
+  color: var(--content-subtle);
+}
