+++ apps/examples/FeedList/shared/refresh-header.css
@@ -1,0 +1,21 @@
+@import "@lynx-js/luna-styles/index.css";
+
+.refresh-header {
+  width: 100%;
+  height: 50px;
+  position: relative;
+  display: flex;
+  flex-direction: row;
+  justify-content: center;
+  align-items: center;
+}
+
+.refresh-header__spinner {
+  width: 50px;
+  height: 50px;
+}
+
+.refresh-header__text {
+  font-size: 14px;
+  color: var(--content-subtle);
+}
